test(opentelemetry-configuration): simplify management of environment variables
The ConfigProvider tests were deleting a number of environment
variables after every test, and this logic was duplicated in a
few places. This is a bit error prone.

The tests seem to pass with a clean process.env provided to each
test, and Mocha's afterEach() hook is inherited, so the logic can
be simplified a good bit while maintaining semantics.