fix(otlp-exporter-base): ensure retry on network errors during HTTP export
OpenTelemetry OTLP/HTTP specification states:

```
If the server disconnects without returning a response, the client SHOULD
retry and send the same request. The client SHOULD implement an exponential
backoff strategy between retries to avoid overwhelming the server.
...
If the client cannot connect to the server, the client SHOULD retry the
connection using an exponential backoff strategy between retries.
The interval between retries must have a random jitter.
```

The backoff infrastrucure was in place, it was just the glue code to
request APIs (fetch, http, xhr) that was reporting non-retryable state
in case of errors that might be temporary.